expert

+++ src/basic/propensity_score_matching.h
@@ -6,7 +6,7 @@ class Propensity_Score_Matching:public sae::Solver<double> {
 public:
     Propensity_Score_Matching(sae::io::MappedGraph *graph);
     ~Propensity_Score_Matching();
-    std::vector<int> solve(int);
+    std::vector<int> solve(double);
     double compare(int);
 
 private:
